extract tx description to a separate component
