+++ src/routes/safe/components/TransactionsNew/TxsTable/ExpandedTx/style.js
@@ -1,7 +1,5 @@
 // @flow
-import {
-  md, lg, connected, error,
-} from '~/theme/variables'
+import { md, lg, connected, error } from '~/theme/variables'
 
 export const styles = () => ({
   txDataContainer: {
